Extracted tag mapping to separate function
